requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
opencv-python==4.8.1.78
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0 